pyproject.toml
[project]
name = "admob-mcp"
version = "0.1.0"
description = "MCP server for AdMob analytics"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
admob-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
